+++ src/main/java/codesquad/springcafe/domain/comment/dto/Comment.java
@@ -1,0 +1,5 @@
+package codesquad.springcafe.domain.comment.dto;
+
+public class Comment {
+
+}
