+++ src/main/java/codesquad/springcafe/domain/user/dto/LoginUser.java
@@ -1,4 +1,4 @@
-package codesquad.springcafe.user.domain;
+package codesquad.springcafe.domain.user.dto;
 
 import java.io.Serializable;
