+++ src/main/java/codesquad/springcafe/domain/user/respository/UserRepository.java
@@ -1,7 +1,7 @@
-package codesquad.springcafe.user.respository;
+package codesquad.springcafe.domain.user.respository;
 
+import codesquad.springcafe.domain.user.dto.User;
 import codesquad.springcafe.exceptions.NoSuchUserException;
-import codesquad.springcafe.user.domain.User;
 
 import java.util.List;
