+++ src/main/java/codesquad/springcafe/domain/user/service/UserServiceImpl.java
@@ -1,11 +1,11 @@
-package codesquad.springcafe.user.service;
+package codesquad.springcafe.domain.user.service;
 
+import codesquad.springcafe.domain.user.dto.LoginUser;
+import codesquad.springcafe.domain.user.dto.User;
+import codesquad.springcafe.domain.user.dto.UserIdentity;
+import codesquad.springcafe.domain.user.respository.UserRepository;
 import codesquad.springcafe.exceptions.CanNotLoginException;
 import codesquad.springcafe.exceptions.NoSuchUserException;
-import codesquad.springcafe.user.domain.LoginUser;
-import codesquad.springcafe.user.domain.User;
-import codesquad.springcafe.user.domain.UserIdentity;
-import codesquad.springcafe.user.respository.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
